+++ src/main/java/se/lexicon/ecommerce/repository/OrderItemRepository.java
@@ -1,13 +1,35 @@
 package se.lexicon.ecommerce.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import se.lexicon.ecommerce.model.OrderItem;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, BigInteger> {
+/**
+ * Repository for {@link OrderItem} persistence and optional order-item queries.
+ */
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 
+    /**
+     * Finds order items by order id.
+     *
+     * @param orderId order id
+     * @return matching order items
+     */
     List<OrderItem> findByOrderId(Long orderId);
-    List<OrderItem> findByProductId(BigInteger productId);
+
+    /**
+     * Finds order items by product id.
+     *
+     * @param productId product id
+     * @return matching order items
+     */
+    List<OrderItem> findByProductId(Long productId);
+
+    /**
+     * Finds order items with quantity greater than the provided value.
+     *
+     * @param quantity quantity threshold
+     * @return matching order items
+     */
     List<OrderItem> findByQuantityGreaterThan(int quantity);
 }
